safari-600.7-branch/LayoutTests/fast/events/popup-blocking-timers5-expected.txt

Click Here
Test calling window.open() with a 1000 ms delay. A popup should be allowed.
PASS newWindow is non-null.
